chore: clean up code and move to HierarchyBuilder
